src/back/backend/requirements/common.txt
asgiref==3.5.0
asttokens==2.0.5
auth==0.5.3
blinker==1.4
boto3==1.21.36
botocore==1.24.36
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-rest-auth==0.9.5
djangorestframework==3.13.1
dnspython==2.2.1
drf-jwt==1.19.2
dynaconf==3.1.7
eventlet==0.33.0
executing==0.8.3
falcon==3.1.0
greenlet==1.1.2
gunicorn==20.1.0
hvac==0.11.2
i==2.1.0
idna==3.3
Jinja2==3.1.1
jmespath==1.0.0
loguru==0.6.0
MarkupSafe==2.1.1
mongoengine==0.24.1
npm==0.1.1
oauthlib==3.2.0
optional-django==0.1.0
Pillow==9.0.1
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pymongo==4.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
validators==0.18.2
Werkzeug==2.1.1